pyproject.toml
[tool.poetry]
name = "interview-task"
version = "0.1.0"
description = ""
authors = ["Kacper Falkiewicz"]
readme = "README.md"
packages = [{include = "interview_task"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
djangorestframework = "^3.14.0"
black = "^23.3.0"
mypy = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
